demo2/lambda-image/requirements.txt
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.3.0
pdf2image==1.17.0
